requirements.txt
Flask
pymongo
diffusers
werkzeug
torch
Pillow